Refactoring sign in screen so that sign in form is in a seperate component.
