+++ .env
@@ -1,0 +1,2 @@
+EXPO_PUBLIC_SUPABASE_URL=https://nzziroohrdngtpurdfxm.supabase.co
+EXPO_PUBLIC_SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im56emlyb29ocmRuZ3RwdXJkZnhtIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NTMxOTgyMDMsImV4cCI6MjA2ODc3NDIwM30.gAqE_HHE_rXFPCE_-2AlrxMQFS4lYGOmrgu6oFzYx1U
