+++ package.json
@@ -17,14 +17,15 @@
     "expo": "^54.0.7",
     "expo-font": "~13.3.2",
     "expo-status-bar": "~2.2.3",
-    "react": "19.0.0",
-    "react-native": "0.79.5",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.11.1"
+    "react": "19.1.0",
+    "react-native": "0.81.4",
+    "react-native-safe-area-context": "~5.6.0",
+    "react-native-screens": "~4.16.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
-    "@types/react": "~19.0.10",
+    "@react-native-community/cli": "latest",
+    "@types/react": "^19.1.13",
     "typescript": "~5.8.3"
   },
   "private": true,
